fix shaders for lighting_demo.cpp as mixing floats with ints for math operators like division isn't supported in webgl. WebGL requires a strict typing consistency otherwise shader code fails to compile. Replace texture image file for textures demo to ensure cmake build pipeline can build all the files successfully. This was due to a missing texture file reference. comment out 3d_model demo from cmake build as I don't have a 3d_models folder containing reference 3d models included locally in this current file system. Use --preload-file instead of --embed-file for the lighting_demo cmakelists file so that I can easily make changes to the generate .data shader file in the build/web/demos/lighting output. Why? because if I wanted to change which shader logic to use I can easily do so without needing to recompile the c++ project. --embed-file embeds the shader code into a wasm file along with the code whereas --preload-file keeps it in a separate file giving me the flexibility to change shader code on the fly while developing on web

+++ shaders/1.colors.vert
@@ -31,7 +31,7 @@ void specularLighting()
     vec3 reflectDir = reflect(-lightDir, norm);
 
     // specular lighting calculation
-    int shininess = 128; // shininess of the highlight - the higher the number the more it reflects the light instead of scattering it all around and makes the highlight smaller
+    float shininess = 128.0; // shininess of the highlight - the higher the number the more it reflects the light instead of scattering it all around and makes the highlight smaller
     float spec = pow(max(dot(viewDir, reflectDir), 0.0), shininess);
     vec3 specular = specularStrength * spec * lightColor;
